export, grouping: move function to script (only used there), clean unused files.
