+++ export.py
@@ -1,0 +1,2 @@
+import runpy
+runpy.run_module('ping.scripts.export', run_name='__main__')
